tests/wpt/web-platform-tests/offscreen-canvas/the-offscreen-canvas/size.attributes.parse.octal.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by tools/gentest.py. -->
<title>OffscreenCanvas test: size.attributes.parse.octal</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/canvas-tests.js"></script>

<h1>size.attributes.parse.octal</h1>
<p class="desc">Parsing of non-negative integers</p>


<script>
var t = async_test("Parsing of non-negative integers");
t.step(function() {

var offscreenCanvas = new OffscreenCanvas(100, 50);
var ctx = offscreenCanvas.getContext('2d');

offscreenCanvas.width = '0100';
offscreenCanvas.height = '0100';
_assertSame(offscreenCanvas.width, 100, "offscreenCanvas.width", "100");
_assertSame(offscreenCanvas.height, 100, "offscreenCanvas.height", "100");

t.done();

});
</script>
