third_party/WebKit/LayoutTests/imported/wpt/offscreen-canvas/the-offscreen-canvas/size.attributes.parse.octal-expected.txt

This is a testharness.js-based test.
FAIL Parsing of non-negative integers assert_equals: offscreenCanvas.height === 100 (got 50[number], expected 100[number]) expected 100 but got 50
Harness: the test ran to completion.
